Texture work: add destructor, fix issue #4, and fix a related memory leak.
This may sound like a lot of work for one commit, but all three problems were
symptoms of the same underlying problem - namely, Texture::Texture(SDL_Surface*,
GLenum) did not copy the surface passed in.  This meant that it could not take
ownership of the surface, and lead to a memory leak in the other constructor
(the other constructor did not free the surface). Additionally, since OpenGL
and common texture formats like PNG have opposite ideas of what down is, images
loaded from a file were vertically flipped.

The new implementation copies the provided surface to a new surface (vertically
flipping it to the correct orientation in the process) and properly frees the
image in Texture::~Texture().  Additionally, I have factored out Texture::init
to fix the memory leak.